requirements.txt
Django==5.1.2
python-dotenv==1.0.1
psycopg2==2.9.10
redis==5.2.0
django-redis==5.4.0
djangorestframework==3.15.2